Add handler for root command and remove the old one
